+++ effectful-core/CHANGELOG.md
@@ -1,2 +1,6 @@
+# effectful-core-1.1.0.0 (2022-07-??)
+* Don't reset the `UnliftStrategy` to `SeqUnlift` inside the continuation of
+  `withEffToIO`.
+
 # effectful-core-1.0.0.0 (2022-07-13)
 * Initial release.